+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Version 0.2.5 (2025-06-19)
+
+- Fix bug with empty lists. Now should parse correctly to empty lists instead of returning defaults.
+## Version 0.2.4 (2025-06-18)
+
+- Added support for SkipJsonSchema fields. They will automatically be excluded from the form and defaults used for validation. 
 ## Version 0.2.3 (2025-06-16 )
 
 - Removed the custom css injection for compact spacing. Instead applying to components directly.